[FRONTEND] Extract AccountsTableBody from AccountsList
